Add files via upload
The program takes a CSV file and extracts/processes data, then stores it into an object. It then creates a dictionary of objects holding the data. 

You can run it by compiling it as you would a regular python program, then providing a single argument in the terminal: the path name. 

The python is excellent for text processing due to its sheer versatility. There are many libraries that can be taken advantage of for processing that aren't really available in Java/C++, especially from a machine learning standpoint. I don't think there are any real weaknesses with python, except maybe its speed, but it isn't that big a deal. 

Mostly everything in this homework was a review because I've done basic text processing as a part of the pre-processing step in a project in CS 4375: intro to machine learning. I'm more interested in the actual application of machine learning algorithms to text, which is something I have not dove deep into.